dashboard/requirements.txt
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.0.0
google-cloud-bigquery>=3.14.0
db-dtypes>=1.2.0
python-dotenv>=1.0.0
pytest>=8.0.0
